requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
pandas
requests